pyproject.toml
[project]
name = "jewel"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=2.0.0",
    "django==4.1.1",
    "django-extensions==3.2.1",
    "django-filter==22.1",
    "djangorestframework==3.14.0",
    "factory-boy>=3.3.0",
    "fuzzywuzzy>=0.18.0",
    "ipdb==0.13.9",
    "ipython==8.5.0",
    "markdown==3.4.1",
    "pandas>=2.3.1",
    "pre-commit==2.20.0",
    "price-parser==0.3.4",
    "psycopg2-binary>=2.9.0",
    "pytest==7.1.3",
    "pytest-django==4.5.2",
    "pytest-watch==4.2.0",
    "requests>=2.32.4",
    "ruff>=0.8.0",
    "tqdm==4.64.1",
]

[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = [
    "E4",      # pycodestyle errors (import)
    "E7",      # pycodestyle errors (statement)
    "E9",      # pycodestyle errors (runtime)
    "F",       # Pyflakes
    "I",       # isort
    "DJ",      # flake8-django
    "UP",      # pyupgrade
    "B",       # flake8-bugbear
    "C4",      # flake8-comprehensions
    "SIM",     # flake8-simplify
]

ignore = [
    "DJ001",   # Django models without __str__ method
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint.isort]
# Django-aware isort configuration
known-first-party = ["jewel", "games", "steam", "csv_importer", "abstract"]
section-order = ["future", "standard-library", "django", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
django = ["django"]
